Plug the memory leak from gss_OID_sap_to_loc
This routine is only used to translate data coming from SAP through
the SNC Adapter into the ABI used by the local GSS-API library.
As such, the output of the routine is never exposed outside of the
shim layer, so we can make all callers (in the shim layer) free its
output, closing the memory leak.  (But don't try to free a NULL OID.)