Plug the memory leak from gss_OID_loc_to_sap
The GSS-API library returns OIDs as pointers to fixed storage, and
the caller is not expected to free them.  However, these OIDs are
in the wrong ABI for SAP, so the shim adaptor must allocate storage
for OIDs in the correct ABI.  To preserve the "caller does not need
to free" property without introducing memory leaks, we must introduce
some form of caching.  The amount of code necessary to do this is
unfortunately large, so we use the simplest conceivable method of
a flat array, extended each time an insertion is necessary.  Locking
around access to the cache is required, to preserve thread-safe behavior.
We do not expect more than 30 OIDs to be necessary, so the O(n) search
behavior is tolerable.